feat: adds utility to check if passed object is iterable
Motivation - Want string to be treated as uniterable, so that
can use logic such as len() on iterables only, without triggering
the iter() of string resulitng in character array from code block
using len() logic.

This provides facility to cover string type as both iterable or
un-iterable. The function provides an input parameter that acts
as flag to treat the string as iterable or otherwise. By default
Python iter() implementation treats string as iterable, which the
user may not always want.