feat: adds functionality for generating md5 checksum
Motivation ==> To provide ability to check for correct file upload to S3
bucket.

The module provides functionality to generate checksum for either a
string or a file. The file can either be read as whole or in parts ie.
in chunks. User can control the chunk size by specifying chunk size in
bytes.

\# EN1-T5